// apps/api/generated/server/model/redeemRequest.ts
/**
 * QZD Platform API
 * The QZD Platform API exposes authentication, account, ledger, remittance, and validator administration capabilities for wallet operators and platform integrations. The contract is defined up-front to enable client/server scaffolding via OpenAPI Generator. 
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RedemptionDestination } from './redemptionDestination.js';


export interface RedeemRequest { 
    fromAccountId: string;
    amount: string;
    destination?: RedemptionDestination;
    memo?: string;
}